requirements.txt
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pymongo
fastapi-mail
redis